requirements.txt
asgiref==3.9.1
Django==4.2.7
djangorestframework==3.14.0
djangorestframework-api-key==3.1.0
gunicorn==21.2.0
packaging==25.0
pytz==2025.2
sqlparse==0.5.3
typing_extensions==4.15.0
